Add quick actions on home screen on mobile
