+++ packages/app/features/home/TokenDetails.tsx
@@ -1,58 +1,19 @@
-import {
-  AnimatePresence,
-  Button,
-  Card,
-  LinkableButton,
-  Paragraph,
-  Separator,
-  Spinner,
-  Stack,
-  Theme,
-  XStack,
-  YStack,
-} from '@my/ui'
-import { type CoinWithBalance, sendCoin, usdcCoin } from 'app/data/coins'
+import { Card, Paragraph, Separator, Spinner, Stack, Theme, XStack, YStack } from '@my/ui'
+import type { CoinWithBalance } from 'app/data/coins'
 import { ArrowDown, ArrowUp } from '@tamagui/lucide-icons'
-import { IconError, IconPlus, IconSwap } from 'app/components/icons'
+import { IconError } from 'app/components/icons'
 import { useTokenMarketData } from 'app/utils/coin-gecko'
 import formatAmount from 'app/utils/formatAmount'
 import { TokenActivity } from './TokenActivity'
 import { useTokenPrices } from 'app/utils/useTokenPrices'
 import { convertBalanceToFiat } from 'app/utils/convertBalanceToUSD'
 import { IconCoin } from 'app/components/icons/IconCoin'
-import { useHoverStyles } from 'app/utils/useHoverStyles'
-
-export function AnimateEnter({ children }: { children: React.ReactNode }) {
-  return (
-    <AnimatePresence>
-      <Stack
-        key="enter"
-        animateOnly={['transform', 'opacity']}
-        animation="200ms"
-        enterStyle={{ opacity: 0, scale: 0.9 }}
-        exitStyle={{ opacity: 0, scale: 0.95 }}
-        opacity={1}
-      >
-        {children}
-      </Stack>
-    </AnimatePresence>
-  )
-}
+import { HomeQuickActions } from 'app/features/home/HomeQuickActions'
 
 export const TokenDetails = ({ coin }: { coin: CoinWithBalance }) => {
-  const hoverStyles = useHoverStyles()
-
-  const getTradeUrl = () => {
-    if (coin?.symbol === sendCoin.symbol) {
-      return `/trade?inToken=${coin?.token}&outToken=${usdcCoin.token}`
-    }
-
-    return `/trade?inToken=${coin?.token}`
-  }
-
   return (
     <YStack f={1} gap="$5" $gtLg={{ w: '45%', pb: '$0' }} pb="$5">
-      <YStack gap="$5">
+      <YStack gap="$3.5" $gtLg={{ gap: '$5' }}>
         <Card p="$4.5" w={'100%'} jc={'space-between'} $gtLg={{ h: 244, p: '$6' }}>
           <YStack gap="$4">
             <XStack ai={'center'} gap={'$3'}>
@@ -74,46 +35,7 @@ export const TokenDetails = ({ coin }: { coin: CoinWithBalance }) => {
             </YStack>
           </YStack>
         </Card>
-        <XStack w={'100%'} gap={'$5'}>
-          <LinkableButton
-            href="/deposit"
-            f={1}
-            height={'auto'}
-            hoverStyle={hoverStyles}
-            focusStyle={hoverStyles}
-          >
-            <YStack gap="$2" jc={'space-between'} ai="center" p="$4">
-              <IconPlus
-                size={'$1.5'}
-                $theme-dark={{ color: '$primary' }}
-                $theme-light={{ color: '$color12' }}
-              />
-              <Button.Text fontSize={'$4'} px="$2">
-                Deposit
-              </Button.Text>
-            </YStack>
-          </LinkableButton>
-          <LinkableButton
-            href={getTradeUrl()}
-            f={1}
-            height={'auto'}
-            hoverStyle={hoverStyles}
-            focusStyle={hoverStyles}
-          >
-            <YStack gap="$2" jc={'space-between'} ai="center" p="$4" height={'auto'}>
-              <Theme name="green">
-                <IconSwap
-                  size={'$1'}
-                  $theme-dark={{ color: '$primary' }}
-                  $theme-light={{ color: '$color12' }}
-                />
-              </Theme>
-              <Button.Text fontSize={'$4'} px="$2">
-                Trade
-              </Button.Text>
-            </YStack>
-          </LinkableButton>
-        </XStack>
+        <HomeQuickActions />
       </YStack>
       <YStack gap={'$3'}>
         <TokenActivity coin={coin} />
